+++ src/components/Home/HorizontalCardList.tsx
@@ -29,8 +29,7 @@ const cardMap = {
 }
 
 const create = require('../../../images/create.png')
-const currentLocation = require('../../../images/currentLocationButton.png')
-const notificationsButton = require('../../../images/notificationsButton.png')
+const mapOptionsButton = require('../../../images/mapOptions.png')
 const height = 115 * ((minHeight - 1) * 0.4 + 1)
 const marginBottom = 14 * s
 const totalHeight = height + marginBottom
@@ -99,27 +98,11 @@ const ButtonColumn = inject('homeStore', 'navStore', 'locationStore', 'wocky')(
           paddingRight: buttonPadding,
         }}
       >
-        <TouchableOpacity
-          onPress={() => {
-            homeStore.stopFollowingUserOnMap()
-            if (locationStore.location) homeStore.setFocusedLocation(locationStore.location)
-          }}
-          style={[styles.button, styles.shadow]}
-        >
-          <Image source={currentLocation} />
-        </TouchableOpacity>
-
         {!homeStore.fullScreenMode && (
           <View>
-            <TouchableOpacity
-              onPress={() => {
-                Actions.notifications()
-              }}
-              style={styles.button}
-            >
+            <TouchableOpacity onPress={Actions.mapOptions} style={styles.button}>
               <View>
-                <Image source={notificationsButton} />
-                {wocky.notifications.hasUnread && <View style={styles.newDot} />}
+                <Image source={mapOptionsButton} />
               </View>
             </TouchableOpacity>